# Improve create query