### Description
This PR improves the create method by moving the logic into one query.

### Query
The create query:
- If there is no previous entry for a given key the query inserts the new row with the key and its value.
- If there is a previous entry for a given key exists:
  -  where the previous entry is not a deleted entry, the query inserts the new row with the key and its value.
  -  where the previous entry is a deleted entry, the query does nothing.

Then we check if any rows have been affected by the query, if not we return a key exists error.

### Index
Note: we will make an update to the `kine_name_index` when we update the schema version to include `delete`. This will allow us to have a covering index for the new create query.