@louiseschmidtgen 2024-08-01
## Benchmark Analysis
The create query takes less (because of the network most likely):
```
                    │    Before    │                After                │
                    │    sec/op    │    sec/op     vs base               │
Create/sqlite-4       1.548m ± 17%   1.443m ± 15%        ~ (p=0.053 n=7)
Create/dqlite-4       1.427m ± 13%   1.272m ± 14%  -10.88% (p=0.026 n=7)
```
The total amount of pages read for the transaction decreased by quite a lot:
```
                    │    Before     │                After                 │
                    │ page-reads/op │ page-reads/op  vs base               │
Create/sqlite-4         21.27 ±  2%    13.92 ±   4%  -34.56% (p=0.001 n=7)
Create/dqlite-4         16.75 ±  3%    10.10 ±   2%  -39.70% (p=0.001 n=7)
```
The transaction is spending way less time reading than before. Consider that this most likely compensates the increase in time spent writing (it spent additional 16µs on the write as you pointed out, but 35µs less during the reads).
```
                    │     Before      │                  After                  │
                    │ sec-reading/op  │ sec-reading/op  vs base                 │
Create/sqlite-4        91.30µ ± 23%      43.70µ ± 879%  -52.14% (p=0.026 n=7)
Create/dqlite-4       44.700µ ± 19%      9.100µ ±  55%  -79.64% (p=0.001 n=7)
```
Allocates way less both in number of allocations and in size:
```
                    │     Before     │                After                 │
                    │      B/op      │     B/op       vs base               │
Create/sqlite-4        23.81Ki ±  0%   19.31Ki ±  2%  -18.90% (p=0.001 n=7)
Create/dqlite-4        21.55Ki ±  0%   17.90Ki ±  0%  -16.91% (p=0.001 n=7)
                    │  allocs/op   │  allocs/op    vs base                 │
Create/sqlite-4        535.0 ±  0%    424.0 ±  4%  -20.75% (p=0.001 n=7)
Create/dqlite-4        397.0 ±  0%    336.0 ±  0%  -15.37% (p=0.001 n=7)
```
We write and read dramatically less from/to the network:
```
                    │        Before         │                    After                     │
                    │ network-bytes-read/op │ network-bytes-read/op  vs base               │
Create/dqlite-4                 423.3 ±  0%             239.4 ±  0%  -43.44% (p=0.001 n=7)
                    │          Before          │                      After                      │
                    │ network-bytes-written/op │ network-bytes-written/op  vs base               │
Create/dqlite-4                  1040.0 ±   0%               313.0 ±   0%  -69.90% (p=0.001 n=7)
```