@louiseschmidtgen 2024-08-01
## Traces
Traces confirm that our k8s cluster is happy and our create query works as expected:
![Screenshot from 2024-08-01 19-09-48](https://github.com/user-attachments/assets/a009aa29-ae2e-4e97-ac8f-5f5e69428df2)