@marco6 2024-08-02
> code-wise LGMT Not sure about the query improvements, I let @marco6 give the final approval on this.
> 
> I see that this decreases read time and amount of reads but wouldn't jump straight to the conclusion that the decrease in reads outweighs the increase in write time. This highly depends on the workload. As said, up to @marco6 to decide.

First of all, I want to point out that the time that increased is the "time spent in a write transaction", which is slightly different from the "time spent writing".

To me, it is a clear improvement. If we look at the performance in Dqlite we get a nice 10% decrease of latency. All resource usage is decreasing sharply. I would accept that even if the latency would increase TBH.

As a side note, this query is "incomplete" in a sense: the query can be made dramatically faster with a slight change in one of the indexes. Most likely, in that case, we could even remove this increase in time spent in a write transaction.

So, green light from my part, but maybe @ktsakalozos has a different opinion? I.e. we might want to decide that it's better to wait for an update in the index?

As a way to set up our minds, we could have a run with kube burner like we were doing in the last iteration so that we understand better and then decide.